requirements.txt
Django==1.11.3
django-extensions==1.8.1
bootstrap-admin==0.3.7.1
djangorestframework==3.6.3
markdown==2.6.8
django-filter==1.0.4
Pygments==2.2.0
django-import-export==0.5.1
django-bootstrap3==9.0.0
Werkzeug==0.12.2
ipython==6.1.0
django-cors-middleware==1.3.1
django-oauth-toolkit==1.0.0
